refactor(knowledge-migration): name the legacy file mappings module in camelCase
Per naming-conventions §3.2, a module that exports only functions takes a
camelCase filename. Rename only, no logic change.

Note that the rest of `mappings/` stays PascalCase deliberately: every file
there is equally function-only, and §2 Core Principle 1 ("consistency beats
style choice", which trumps the §3 tables) says a directory that
consistently uses one convention is healthier than one mixing two. Making
that directory uniformly camelCase is a separate change — `PreferencesMappings.ts`
and `BootConfigMappings.ts` are generated by the data-classify pipeline, so
it would have to go through the generator.

Reported in review of #18167.
